requirements.txt
requests>=2.26.0
loguru>=0.5.3
retrying>=1.3.3
pymysql>=1.0.2
dbutils>=2.0.2
glom>=20.11.0
redis>=3.5.3
furl>=2.1.2
arrow>=1.1.1
pytz>=2021.1
grpcio-tools>=1.39.0
kombu>=5.1.0
croniter>=1.0.15
schedule>=1.1.0
cryptography>=3.4.8
lxml>=4.6.3
stringcase>=1.2.0
sqlalchemy>=1.4.23
pymongo>=3.12.0
amqp>==5.0.6